+++ src/Vod/V20170321/VodApiResolver.php
@@ -18,13 +18,16 @@ use AlibabaCloud\Client\Resolver\ApiResolver;
  * @method BatchStopVodDomain batchStopVodDomain(array $options = [])
  * @method CreateAppInfo createAppInfo(array $options = [])
  * @method CreateAudit createAudit(array $options = [])
+ * @method CreateMediaDNALib createMediaDNALib(array $options = [])
  * @method CreateUploadAttachedMedia createUploadAttachedMedia(array $options = [])
  * @method CreateUploadImage createUploadImage(array $options = [])
  * @method CreateUploadVideo createUploadVideo(array $options = [])
+ * @method DeleteAIImageInfos deleteAIImageInfos(array $options = [])
  * @method DeleteAITemplate deleteAITemplate(array $options = [])
  * @method DeleteAppInfo deleteAppInfo(array $options = [])
  * @method DeleteAttachedMedia deleteAttachedMedia(array $options = [])
  * @method DeleteCategory deleteCategory(array $options = [])
+ * @method DeleteDynamicImage deleteDynamicImage(array $options = [])
  * @method DeleteEditingProject deleteEditingProject(array $options = [])
  * @method DeleteImage deleteImage(array $options = [])
  * @method DeleteMessageCallback deleteMessageCallback(array $options = [])
@@ -53,9 +56,12 @@ use AlibabaCloud\Client\Resolver\ApiResolver;
  * @method DescribeVodRefreshQuota describeVodRefreshQuota(array $options = [])
  * @method DescribeVodRefreshTasks describeVodRefreshTasks(array $options = [])
  * @method DescribeVodStorageData describeVodStorageData(array $options = [])
+ * @method DescribeVodTagResources describeVodTagResources(array $options = [])
  * @method DescribeVodTranscodeData describeVodTranscodeData(array $options = [])
  * @method DescribeVodUserDomains describeVodUserDomains(array $options = [])
+ * @method DescribeVodUserTags describeVodUserTags(array $options = [])
  * @method DetachAppPolicyFromIdentity detachAppPolicyFromIdentity(array $options = [])
+ * @method GetAIImageJobs getAIImageJobs(array $options = [])
  * @method GetAIMediaAuditJob getAIMediaAuditJob(array $options = [])
  * @method GetAITemplate getAITemplate(array $options = [])
  * @method GetAIVideoTagResult getAIVideoTagResult(array $options = [])
@@ -64,6 +70,8 @@ use AlibabaCloud\Client\Resolver\ApiResolver;
  * @method GetAuditHistory getAuditHistory(array $options = [])
  * @method GetCategories getCategories(array $options = [])
  * @method GetDefaultAITemplate getDefaultAITemplate(array $options = [])
+ * @method GetDRMCertInfo getDRMCertInfo(array $options = [])
+ * @method GetDRMLicense getDRMLicense(array $options = [])
  * @method GetEditingProject getEditingProject(array $options = [])
  * @method GetEditingProjectMaterials getEditingProjectMaterials(array $options = [])
  * @method GetImageInfo getImageInfo(array $options = [])
@@ -86,13 +94,16 @@ use AlibabaCloud\Client\Resolver\ApiResolver;
  * @method GetVideoPlayAuth getVideoPlayAuth(array $options = [])
  * @method GetVodTemplate getVodTemplate(array $options = [])
  * @method GetWatermark getWatermark(array $options = [])
+ * @method ListAIImageInfo listAIImageInfo(array $options = [])
  * @method ListAIJob listAIJob(array $options = [])
  * @method ListAITemplate listAITemplate(array $options = [])
  * @method ListAppInfo listAppInfo(array $options = [])
  * @method ListAppPoliciesForIdentity listAppPoliciesForIdentity(array $options = [])
  * @method ListAuditSecurityIp listAuditSecurityIp(array $options = [])
+ * @method ListDynamicImage listDynamicImage(array $options = [])
  * @method ListLiveRecordVideo listLiveRecordVideo(array $options = [])
  * @method ListMediaDNADeleteJob listMediaDNADeleteJob(array $options = [])
+ * @method ListMediaDNALibs listMediaDNALibs(array $options = [])
  * @method ListSnapshots listSnapshots(array $options = [])
  * @method ListTranscodeTask listTranscodeTask(array $options = [])
  * @method ListTranscodeTemplateGroup listTranscodeTemplateGroup(array $options = [])
@@ -113,6 +124,7 @@ use AlibabaCloud\Client\Resolver\ApiResolver;
  * @method SetEditingProjectMaterials setEditingProjectMaterials(array $options = [])
  * @method SetMessageCallback setMessageCallback(array $options = [])
  * @method SetVodDomainCertificate setVodDomainCertificate(array $options = [])
+ * @method SubmitAIImageJob submitAIImageJob(array $options = [])
  * @method SubmitAIJob submitAIJob(array $options = [])
  * @method SubmitAIMediaAuditJob submitAIMediaAuditJob(array $options = [])
  * @method SubmitDynamicImageJob submitDynamicImageJob(array $options = [])
@@ -121,6 +133,8 @@ use AlibabaCloud\Client\Resolver\ApiResolver;
  * @method SubmitSnapshotJob submitSnapshotJob(array $options = [])
  * @method SubmitTranscodeJobs submitTranscodeJobs(array $options = [])
  * @method SubmitWorkflowJob submitWorkflowJob(array $options = [])
+ * @method TagVodResources tagVodResources(array $options = [])
+ * @method UnTagVodResources unTagVodResources(array $options = [])
  * @method UpdateAITemplate updateAITemplate(array $options = [])
  * @method UpdateAppInfo updateAppInfo(array $options = [])
  * @method UpdateAttachedMediaInfos updateAttachedMediaInfos(array $options = [])
@@ -389,6 +403,24 @@ class CreateAudit extends Rpc
 /**
  * @method string getResourceOwnerId()
  * @method $this withResourceOwnerId($value)
+ * @method string getModelType()
+ * @method $this withModelType($value)
+ * @method string getLibRegion()
+ * @method $this withLibRegion($value)
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
+ * @method string getOwnerAccount()
+ * @method $this withOwnerAccount($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ */
+class CreateMediaDNALib extends Rpc
+{
+}
+
+/**
+ * @method string getResourceOwnerId()
+ * @method $this withResourceOwnerId($value)
  * @method string getIcon()
  * @method $this withIcon($value)
  * @method string getDescription()
@@ -505,6 +537,20 @@ class CreateUploadVideo extends Rpc
  * @method $this withResourceOwnerAccount($value)
  * @method string getOwnerId()
  * @method $this withOwnerId($value)
+ * @method string getAIImageInfoIds()
+ * @method $this withAIImageInfoIds($value)
+ */
+class DeleteAIImageInfos extends Rpc
+{
+}
+
+/**
+ * @method string getResourceOwnerId()
+ * @method $this withResourceOwnerId($value)
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
  * @method string getTemplateId()
  * @method $this withTemplateId($value)
  */
@@ -553,6 +599,22 @@ class DeleteAttachedMedia extends Rpc
  * @method $this withCateId($value)
  */
 class DeleteCategory extends Rpc
+{
+}
+
+/**
+ * @method string getResourceOwnerId()
+ * @method $this withResourceOwnerId($value)
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
+ * @method string getVideoId()
+ * @method $this withVideoId($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ * @method string getDynamicImageIds()
+ * @method $this withDynamicImageIds($value)
+ */
+class DeleteDynamicImage extends Rpc
 {
 }
 
@@ -1007,6 +1069,55 @@ class DescribeVodStorageData extends Rpc
 }
 
 /**
+ * @method string getScope()
+ * @method $this withScope($value)
+ * @method array getTag()
+ * @method array getResourceId()
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ * @method string getResourceType()
+ * @method $this withResourceType($value)
+ */
+class DescribeVodTagResources extends Rpc
+{
+
+    /**
+     * @param array $tag
+     *
+     * @return $this
+     */
+	public function withTag(array $tag)
+	{
+	    $this->data['Tag'] = $tag;
+		foreach ($tag as $depth1 => $depth1Value) {
+			if(isset($depth1Value['Key'])){
+				$this->options['query']['Tag.' . ($depth1 + 1) . '.Key'] = $depth1Value['Key'];
+			}
+			if(isset($depth1Value['Value'])){
+				$this->options['query']['Tag.' . ($depth1 + 1) . '.Value'] = $depth1Value['Value'];
+			}
+		}
+
+		return $this;
+    }
+
+    /**
+     * @param array $resourceId
+     *
+     * @return $this
+     */
+	public function withResourceId(array $resourceId)
+	{
+	    $this->data['ResourceId'] = $resourceId;
+		foreach ($resourceId as $i => $iValue) {
+			$this->options['query']['ResourceId.' . ($i + 1)] = $iValue;
+		}
+
+		return $this;
+    }
+}
+
+/**
  * @method string getStartTime()
  * @method $this withStartTime($value)
  * @method string getStorage()
@@ -1037,6 +1148,7 @@ class DescribeVodTranscodeData extends Rpc
  * @method $this withCdnType($value)
  * @method string getPageSize()
  * @method $this withPageSize($value)
+ * @method array getTag()
  * @method string getFuncFilter()
  * @method $this withFuncFilter($value)
  * @method string getDomainName()
@@ -1051,6 +1163,34 @@ class DescribeVodTranscodeData extends Rpc
  * @method $this withDomainSearchType($value)
  */
 class DescribeVodUserDomains extends Rpc
+{
+
+    /**
+     * @param array $tag
+     *
+     * @return $this
+     */
+	public function withTag(array $tag)
+	{
+	    $this->data['Tag'] = $tag;
+		foreach ($tag as $depth1 => $depth1Value) {
+			if(isset($depth1Value['Value'])){
+				$this->options['query']['Tag.' . ($depth1 + 1) . '.Value'] = $depth1Value['Value'];
+			}
+			if(isset($depth1Value['Key'])){
+				$this->options['query']['Tag.' . ($depth1 + 1) . '.Key'] = $depth1Value['Key'];
+			}
+		}
+
+		return $this;
+    }
+}
+
+/**
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ */
+class DescribeVodUserTags extends Rpc
 {
 }
 
@@ -1071,6 +1211,22 @@ class DescribeVodUserDomains extends Rpc
  * @method $this withAppId($value)
  */
 class DetachAppPolicyFromIdentity extends Rpc
+{
+}
+
+/**
+ * @method string getResourceOwnerId()
+ * @method $this withResourceOwnerId($value)
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
+ * @method string getOwnerAccount()
+ * @method $this withOwnerAccount($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ * @method string getJobIds()
+ * @method $this withJobIds($value)
+ */
+class GetAIImageJobs extends Rpc
 {
 }
 
@@ -1201,6 +1357,38 @@ class GetCategories extends Rpc
  * @method $this withTemplateType($value)
  */
 class GetDefaultAITemplate extends Rpc
+{
+}
+
+/**
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
+ * @method string getVideoId()
+ * @method $this withVideoId($value)
+ * @method string getCertId()
+ * @method $this withCertId($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ */
+class GetDRMCertInfo extends Rpc
+{
+}
+
+/**
+ * @method string getCDMData()
+ * @method $this withCDMData($value)
+ * @method string getCertId()
+ * @method $this withCertId($value)
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
+ * @method string getVideoId()
+ * @method $this withVideoId($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ * @method string getDRMType()
+ * @method $this withDRMType($value)
+ */
+class GetDRMLicense extends Rpc
 {
 }
 
@@ -1607,6 +1795,20 @@ class GetWatermark extends Rpc
  * @method $this withResourceOwnerId($value)
  * @method string getResourceOwnerAccount()
  * @method $this withResourceOwnerAccount($value)
+ * @method string getVideoId()
+ * @method $this withVideoId($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ */
+class ListAIImageInfo extends Rpc
+{
+}
+
+/**
+ * @method string getResourceOwnerId()
+ * @method $this withResourceOwnerId($value)
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
  * @method string getOwnerAccount()
  * @method $this withOwnerAccount($value)
  * @method string getOwnerId()
@@ -1681,6 +1883,20 @@ class ListAuditSecurityIp extends Rpc
 /**
  * @method string getResourceOwnerId()
  * @method $this withResourceOwnerId($value)
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
+ * @method string getVideoId()
+ * @method $this withVideoId($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ */
+class ListDynamicImage extends Rpc
+{
+}
+
+/**
+ * @method string getResourceOwnerId()
+ * @method $this withResourceOwnerId($value)
  * @method string getStartTime()
  * @method $this withStartTime($value)
  * @method string getAppName()
@@ -1721,6 +1937,22 @@ class ListLiveRecordVideo extends Rpc
  * @method $this withJobIds($value)
  */
 class ListMediaDNADeleteJob extends Rpc
+{
+}
+
+/**
+ * @method string getResourceOwnerId()
+ * @method $this withResourceOwnerId($value)
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
+ * @method string getOwnerAccount()
+ * @method $this withOwnerAccount($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ * @method string getLibRegion()
+ * @method $this withLibRegion($value)
+ */
+class ListMediaDNALibs extends Rpc
 {
 }
 
@@ -2117,6 +2349,28 @@ class SetVodDomainCertificate extends Rpc
 /**
  * @method string getResourceOwnerId()
  * @method $this withResourceOwnerId($value)
+ * @method string getAIPipelineId()
+ * @method $this withAIPipelineId($value)
+ * @method string getUserData()
+ * @method $this withUserData($value)
+ * @method string getResourceOwnerAccount()
+ * @method $this withResourceOwnerAccount($value)
+ * @method string getOwnerAccount()
+ * @method $this withOwnerAccount($value)
+ * @method string getVideoId()
+ * @method $this withVideoId($value)
+ * @method string getAITemplateId()
+ * @method $this withAITemplateId($value)
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ */
+class SubmitAIImageJob extends Rpc
+{
+}
+
+/**
+ * @method string getResourceOwnerId()
+ * @method $this withResourceOwnerId($value)
  * @method string getUserData()
  * @method $this withUserData($value)
  * @method string getTypes()
@@ -2151,6 +2405,8 @@ class SubmitAIJob extends Rpc
  * @method $this withMediaId($value)
  * @method string getTemplateId()
  * @method $this withTemplateId($value)
+ * @method string getMediaAuditConfiguration()
+ * @method $this withMediaAuditConfiguration($value)
  * @method string getMediaType()
  * @method $this withMediaType($value)
  */
@@ -2286,6 +2542,95 @@ class SubmitTranscodeJobs extends Rpc
  */
 class SubmitWorkflowJob extends Rpc
 {
+}
+
+/**
+ * @method array getTag()
+ * @method array getResourceId()
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ * @method string getResourceType()
+ * @method $this withResourceType($value)
+ */
+class TagVodResources extends Rpc
+{
+
+    /**
+     * @param array $tag
+     *
+     * @return $this
+     */
+	public function withTag(array $tag)
+	{
+	    $this->data['Tag'] = $tag;
+		foreach ($tag as $depth1 => $depth1Value) {
+			if(isset($depth1Value['Key'])){
+				$this->options['query']['Tag.' . ($depth1 + 1) . '.Key'] = $depth1Value['Key'];
+			}
+			if(isset($depth1Value['Value'])){
+				$this->options['query']['Tag.' . ($depth1 + 1) . '.Value'] = $depth1Value['Value'];
+			}
+		}
+
+		return $this;
+    }
+
+    /**
+     * @param array $resourceId
+     *
+     * @return $this
+     */
+	public function withResourceId(array $resourceId)
+	{
+	    $this->data['ResourceId'] = $resourceId;
+		foreach ($resourceId as $i => $iValue) {
+			$this->options['query']['ResourceId.' . ($i + 1)] = $iValue;
+		}
+
+		return $this;
+    }
+}
+
+/**
+ * @method array getResourceId()
+ * @method string getOwnerId()
+ * @method $this withOwnerId($value)
+ * @method string getResourceType()
+ * @method $this withResourceType($value)
+ * @method array getTagKey()
+ */
+class UnTagVodResources extends Rpc
+{
+
+    /**
+     * @param array $resourceId
+     *
+     * @return $this
+     */
+	public function withResourceId(array $resourceId)
+	{
+	    $this->data['ResourceId'] = $resourceId;
+		foreach ($resourceId as $i => $iValue) {
+			$this->options['query']['ResourceId.' . ($i + 1)] = $iValue;
+		}
+
+		return $this;
+    }
+
+    /**
+     * @param array $tagKey
+     *
+     * @return $this
+     */
+	public function withTagKey(array $tagKey)
+	{
+	    $this->data['TagKey'] = $tagKey;
+		foreach ($tagKey as $i => $iValue) {
+			$this->options['query']['TagKey.' . ($i + 1)] = $iValue;
+		}
+
+		return $this;
+    }
 }
 
 /**